Use assembly name instead of filename when installing assemblies with gacutil.
gacutil just used the filename (which can be different compared to
the assembly name) in order to install assemblies, though afterwards
entirely relied on the filename to be equal to the assembly name.
In other words, installing assemblies where the filename differed
from the assembly name resulted in an incorrect install: the
files were not visible in the GAC.

Example: gacutil -i mysql.data.dll
Wouldn't return any errors, but the assembly (which is called MySql.Data)
still couldn't be reached.

This patch looks up the assembly name during the installation process and
uses it in order to install an assembly into the GAC:
"gacutil -i mysql.data.dll" will create MySql.Data.dll in the GAC.